nested search, profile, , and * routes under client-portal. Added clientSearch component. Added outlet to clientPortal component.
