+++ outreach-mapping-solutions-fe-react/src/clientProfile.jsx
@@ -1,0 +1,7 @@
+export default function ClientProfile() {
+  return (
+    <div>
+      <p>Client Profile</p>
+    </div>
+  );
+}
